pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cbz"
version = "4.0.0"
description = "Modern Python library for creating, manipulating and viewing comics in CBZ, CBR and PDF formats."
license = "MIT"
authors = ["hyugogirubato <65763543+hyugogirubato@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/hyugogirubato/cbz"
keywords = [
    "python",
    "cbz",
    "cbr",
    "pdf",
    "ebooks",
    "manga",
    "comics",
    "webtoons"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
include = [
    { path = "CHANGELOG.md", format = "sdist" },
    { path = "README.md", format = "sdist" },
    { path = "LICENSE", format = "sdist" },
]

[tool.poetry.urls]
"Issues" = "https://github.com/hyugogirubato/cbz/issues"
"Changelog" = "https://github.com/hyugogirubato/cbz/blob/main/CHANGELOG.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langcodes = ">=3.4.0"
Pillow = ">=10.4.0"
pypdf = ">=5.7.0"
rarfile = ">=4.2"
xmltodict = ">=0.14.2"
pillow-avif-plugin = { version = ">=1.5.2", optional = true }
pillow-jxl-plugin = { version = ">=1.3.4", optional = true }

[tool.poetry.extras]
pillow = ["pillow-avif-plugin", "pillow-jxl-plugin"]

[tool.poetry.scripts]
cbzplayer = "cbz.__main__:main"

[[tool.poetry.source]]
name = "localpypi"
url = "https://pypi.org/simple/"
priority = "primary"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0,<10.0.0"
pytest-cov = ">=5.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=cbz",
]

[certificates]
localpypi = { cert = false }
